+++ library/ruler/src/main/java/io/mehow/ruler/LengthUnit.kt
@@ -2,7 +2,7 @@ package io.mehow.ruler
 
 import java.math.BigDecimal
 
-public interface LengthUnit<T> : Comparable<T>, Iterable<T> where T : LengthUnit<T>, T : Enum<T> {
+public interface LengthUnit<T> : Comparable<T> where T : LengthUnit<T>, T : Enum<T> {
   public val meterRatio: BigDecimal
 
   public operator fun contains(meters: BigDecimal): Boolean