+++ docs/changelog.md
@@ -14,7 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `Distance` and `Length` throw now `ArithmeticException` in case of failures due to overflows or math operations.
 - Rename `exactTotalMeters` property to `meters`.
+- `LengthUnit` interface no longer extends `Iterable`.
 - Upgrade to Kotlin `1.4.21`.
+
+### Removed
+- Companion objects of `SiLengthUnit` and `ImperialLengthUnit`.
 
 ### Fixed
 - Wrong lower meter bound of `Miles` unit.